requirements.txt
fastapi
uvicorn
streamlit
pandas
matplotlib
pydantic
sqlalchemy
python-dotenv
tweepy
transformers
torch
scikit-learn
requests
pytest
